+++ react-native.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  project: {
+    android: {
+      sourceDir: './apps/mobile/android',
+      packageName: 'com.trakt.mobile',
+    },
+  },
+};